+++ tests/test_skoolhtml.py
@@ -1471,7 +1471,7 @@ class SkoolMacroTest(HtmlWriterTestCase, CommonSkoolMacroTest):
 
     def _test_audio_macro(self, writer, macro, src, path=None, delays=None, contention=False,
                           interrupts=False, offset=0, ma_filter=False, is128k=False, config=None,
-                          audio_log=None):
+                          audio_log=None, beeper=0):
         exp_html = f"""
             <audio controls="" src="{src}">
             <p>Your browser doesn't support HTML5 audio. Here is a <a href="{src}">link to the audio</a> instead.</p>
@@ -1504,7 +1504,7 @@ class SkoolMacroTest(HtmlWriterTestCase, CommonSkoolMacroTest):
             self.assertEqual(audio_log, ay_audio_writer.audio_log)
             self.assertEqual(ay_audio_writer.write_options.volume, 75)
             self.assertEqual(ay_audio_writer.write_options.ay_res, 70908)
-            self.assertFalse(ay_audio_writer.write_options.beeper)
+            self.assertIs(ay_audio_writer.write_options.beeper, beeper)
         is128k = len(writer.snapshot) == 0x20000
         aw_config = {
             'ClockSpeed': 3546900 if is128k else 3500000,
@@ -1804,6 +1804,64 @@ class SkoolMacroTest(HtmlWriterTestCase, CommonSkoolMacroTest):
         ]
         self._test_audio_macro(writer, macro, exp_src, exp_path, is128k=True, audio_log=exp_audio_log)
 
+    def test_macro_audio_with_ay_and_beeper(self):
+        skool = """
+            @bank=6
+            ; AY activity with beeper
+            c$C000 LD C,$FD
+             $C002 LD DE,$000F
+            *$C005 LD B,$FF
+             $C007 OUT (C),E
+             $C009 LD B,$BF
+             $C00B OUT (C),D
+             $C00D OUT ($FE),A
+             $C00F XOR $10
+             $C011 INC D
+             $C012 DEC E
+             $C013 JP P,$C005
+             $C016 RET
+        """
+        writer = self._get_writer(skool=skool, mock_file_info=True)
+        fname = 'ay-beeper.wav'
+        macro = f'#AUDIO1,$C000,$C016,,,,,1,1({fname})'
+        exp_src = f'../audio/{fname}'
+        exp_path = f'audio/{fname}'
+        exp_audio_log = [
+            (43, 15, 0),
+            (55, 255, 0),
+            (117, 14, 1),
+            (129, 255, 0),
+            (191, 13, 2),
+            (203, 255, 0),
+            (265, 12, 3),
+            (277, 255, 0),
+            (339, 11, 4),
+            (351, 255, 0),
+            (413, 10, 5),
+            (425, 255, 0),
+            (487, 9, 6),
+            (499, 255, 0),
+            (561, 8, 7),
+            (573, 255, 0),
+            (635, 7, 8),
+            (647, 255, 0),
+            (709, 6, 9),
+            (721, 255, 0),
+            (783, 5, 10),
+            (795, 255, 0),
+            (857, 4, 11),
+            (869, 255, 0),
+            (931, 3, 12),
+            (943, 255, 0),
+            (1005, 2, 13),
+            (1017, 255, 0),
+            (1079, 1, 14),
+            (1091, 255, 0),
+            (1153, 0, 15),
+            (1165, 255, 0),
+        ]
+        self._test_audio_macro(writer, macro, exp_src, exp_path, is128k=True, audio_log=exp_audio_log, beeper=1)
+
     def test_macro_audio_with_contention(self):
         writer = self._get_writer(skool='', mock_file_info=True)
         fname = 'sound.wav'
@@ -1899,7 +1957,7 @@ class SkoolMacroTest(HtmlWriterTestCase, CommonSkoolMacroTest):
         writer = self._get_writer(skool=skool, mock_file_info=True)
         offset = 14335
         fname = 'sound.wav'
-        macro = f'#AUDIO(execint=0,cmio=1,ay=0,offset={offset},maf=0,sim=1,start=24576,stop=24589)({fname})'
+        macro = f'#AUDIO(execint=0,cmio=1,ay=0,bpr=0,offset={offset},maf=0,sim=1,start=24576,stop=24589)({fname})'
         exp_src = f'../audio/{fname}'
         exp_path = f'audio/{fname}'
         exp_delays = [203, 214, 217, 215, 220]